+++ AssemblyInfo.cs
@@ -1,14 +1,14 @@
 ﻿using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("JustSleep")]
+[assembly: AssemblyTitle(JustSleep.JustSleep.pluginName)]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("JustSleep")]
+[assembly: AssemblyProduct(JustSleep.JustSleep.pluginName)]
 [assembly: AssemblyCopyright("Copyright ©  2024")]
 [assembly: AssemblyTrademark("")]
 [assembly: ComVisible(false)]
 [assembly: Guid("1f044c3f-2bd7-41a4-981b-5e0e47ba84e3")]
-[assembly: AssemblyFileVersion("1.0.3")]
-[assembly: AssemblyVersion("1.0.3")]
+[assembly: AssemblyFileVersion(JustSleep.JustSleep.pluginVersion)]
+[assembly: AssemblyVersion(JustSleep.JustSleep.pluginVersion)]